Rename participant previewer component to participant view
