+++ frontend/src/components/participant_view/cohort_landing.scss
