+++ frontend/src/components/participant_view/participant_header.ts
