+++ frontend/src/components/participant_view/participant_nav.ts
