+++ frontend/src/components/participant_view/participant_view.scss
